backend/src/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
qdrant-client>=1.7.1
duckduckgo-search>=4.0.0
langchain-google-genai
langchain-core
langchain-community
langchain-qdrant
langchain-text-splitters
httpx>=0.25.0